// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-outgoing-email-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface AdminOutgoingEmailHttpResponse {
    attemptCount?: number;
    contentHidden?: boolean;
    correlationId?: string | null;
    createdAtUtc?: string;
    fromAddress?: string;
    id?: string;
    inReplyTo?: string | null;
    maxAttempts?: number;
    purpose?: string;
    replyTo?: string | null;
    sentAtUtc?: string | null;
    status?: string;
    subject?: string;
    textBody?: string | null;
    to?: Array<string>;
}