// FoodDiary.Web.Client/projects/fooddiary-admin/src/app/shared/api/sdk/generated/model/admin-outgoing-email-page-http-response.ts
/**
 * FoodDiary Admin API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { AdminOutgoingEmailHttpResponse } from './admin-outgoing-email-http-response';

export interface AdminOutgoingEmailPageHttpResponse {
    items?: Array<AdminOutgoingEmailHttpResponse>;
    statusCounts?: { [key: string]: number } | null;
    totalItems?: number;
}